refactor: AnalysisController에서 필요한 응답의 자료형과 메서드 이름 변경
- 응답 자료형은 최대한 다루기 익숙하고 메모리 크기가 작은 형으로 바꿈 (long -> int)
- 전체 코드 내 통일성을 위해 ID값은 Long형으로 받도록 함
- 단순 getter가 아닌 메서드의 이름에서 get을 쓰지 않도록 수정

+++ src/main/java/com/saebyeok/saebyeok/dto/ArticlesAnalysisResponse.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Getter
 public class ArticlesAnalysisResponse {
-    private long[] articleEmotionsCount;
+    private int[] articleEmotionsCount;
     private String articleAnalysisMessage;
 }
